% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateExtractHead.R
\docType{package}
\name{climateExtract}
\alias{climateExtract}
\title{climateExtract: Extract Climate Data From a Local or online NETCDF}
\description{
A toolbox to ease the extraction and manipulaton of daily gridded climate 
data from a NETCDF object available from the European Climate Assessement & Datasets (ECAD) 
for specific time period, compute summarized values and link them to set of spatial points.
}
\author{
Reto Schmucki
}
