% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingDegreeDay.R
\name{cumsum_rb}
\alias{cumsum_rb}
\title{cumsum_rb}
\usage{
cumsum_rb(x, indices = NULL)
}
\arguments{
\item{x}{rasterbrick object on which cumulative sum should be computed}

\item{indices}{vector with indices over which the sum is to be cumulated.}
}
\description{
Compute the cumulative sum on a multilayered raster (rasterbrick), using indices spaning over specific time periods.
}
\details{
This function computes the cumulative sum over specific periods defined by the a vector of incides (e.g.
two five-day cumulative sum c(1,1,1,1,1,2,2,2,2,2), restarting at zero on the first day of each series defined by
the indices).
}
\author{
Reto Schmucki
}
