% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{ecad_version}
\alias{ecad_version}
\title{version of ECA&D to use}
\format{
character string (e.g. "29.0")
\describe{
  \item{ecad_version}{version number used to gather ECAD data}
}
}
\usage{
ecad_version
}
\description{
version of ECA&D to use
}
\keyword{datasets}
