% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_lc_data_extract.R
\name{esri_lc_data_extract}
\alias{esri_lc_data_extract}
\title{CHIRPS data download.}
\usage{
esri_lc_data_extract(year, utm_id, path)
}
\arguments{
\item{year}{Year.}

\item{utm_id}{UTM ID.}

\item{path}{path in which downloaded data will be stored.}
}
\value{
\code{chirps_data_extract()} download all the GEOTIFF files for the period between year_start and year_end from https://data.chc.ucsb.edu/products/CHIRPS-2.0/global_monthly/tifs/.
}
\description{
CHIRPS data download.
}
\examples{
\dontrun{
esri_lc_data_extract(year,utm_id,path)
}
}
