% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{extract_nc_value}
\alias{extract_nc_value}
\title{extract_nc_value}
\usage{
extract_nc_value(first.year = NULL, last.year = NULL, local_file = TRUE,
  clim_variable = "mean temp", grid_size = 0.5)
}
\arguments{
\item{last.year}{a numeric value defining the last year of the time-period to extract, 2014 if NULL, default=NULL}

\item{local_file}{logical if the .nc data are available on your local disc, if FALSE the data will be downloaded, default=TRUE}

\item{clim_variable}{string defining the daily climate variable of interest; "mean temp","max temp","min temp","precipitation", default="mean temp"}

\item{grid_size}{numeric value in degree defining the resolution of the grid, 0.25 (ca. xx meters) or 0.50 (ca. xx meters), default=0.50}

\item{firs.year}{a numeric value defining the first year of the time-period to extract, 1950 if NULL, default=NULL}
}
\description{
This function extract climate data from NETCDF file produced by the European Climate Assesment & Dataset for a specific time-period and available at http://www.ecad.eu/download/ensembles/download.php#datafiles
}
\details{
This function ask you to select the .nc file containing the data of interest on your local disc, if local_file is FALSE, data will be downloaded from the ECAD. If first.year and last.year are not provided, the function extract the full data set
}
\author{
Reto Schmucki
}

