% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{extrat_nc_value}
\alias{extrat_nc_value}
\title{extract_nc_value}
\usage{
extrat_nc_value(first.year = NULL, last.year = NULL)
}
\arguments{
\item{last.year}{a numeric value defining the last year of the time-period to extract}

\item{firs.year}{a numeric value defining the first year of the time-period to extract}
}
\description{
This function extract climate data from NETCDF file produced by the European Climate Assesment & Dataset for a specific time-period and available at http://www.ecad.eu/download/ensembles/download.php#datafiles
}
\details{
This function ask you to select the .nc file containing the data of interest. If first.year and last.year are not provided, the function extract the full data set
}
\author{
Reto Schmucki
}

