% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingDegreeDay.R
\name{gdd_avg_r}
\alias{gdd_avg_r}
\title{gdd_avg_r}
\usage{
gdd_avg_r(base_temp, min_temp, max_temp, avg_temp, top_temp)
}
\arguments{
\item{base_temp}{value for the base temperature to use for the growing degree day, default = 4°C}

\item{min_temp}{raster, rasterbrick or vector of the minimum daily temperature}

\item{max_temp}{raster, rasterbrick or vector of the maximum daily temperature}

\item{avg_temp}{raster, rasterbrick or vector of the mean daily temperature}

\item{top_temp}{value for the maximum temperature beyond growing degree day are not accumulating, default = NULL}
}
\description{
Compute the growing degree day (GDD) using a base temperature against the mean daily temperature.
}
\details{
This function computes a rough estimate of GDD from the mean temperature without accounting from variation
during the day. For a more refined method using a daily curve, use the BE method (Baskerville-Emin method) available
with `gdd_be_r()`.
}
\author{
Reto Schmucki
}
