% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingDegreeDay.R
\name{gdd_be_r}
\alias{gdd_be_r}
\title{gdd_be_r}
\usage{
gdd_be_r(base_temp, min_temp, max_temp, avg_temp, top_temp)
}
\arguments{
\item{base_temp}{value for the base temperature to use for the growing degree day, default = 4°C}

\item{min_temp}{raster, rasterbrick or vector of the minimum daily temperature}

\item{max_temp}{raster, rasterbrick or vector of the maximum daily temperature}

\item{avg_temp}{raster, rasterbrick or vector of the mean daily temperature}

\item{top_temp}{value for the maximum temperature beyond growing degree day are not accumulating, default = NULL}
}
\description{
Compute the growing degree day (GDD) using the BE method (Baskerville-Emin method) with base temperature.
}
\details{
This function computes an estimate of GDD fitting a sine curve on the minimum and the maximum to account for
daily variation in temperature.
}
\author{
Reto Schmucki
}
