% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingDegreeDay.R
\name{gdd_extract}
\alias{gdd_extract}
\title{ggd_extract
Compute the growing degree day (GDD) using either the mean temperature or the Baskerville-Emin method,
with base temperature.}
\usage{
gdd_extract(
  base_temp = NULL,
  min_temp = NULL,
  max_temp = NULL,
  avg_temp = NULL,
  gdd_method = "avg",
  top_temp = NULL
)
}
\arguments{
\item{base_temp}{value for the base temperature to use for the growing degree day, default = 4°C}

\item{min_temp}{raster, rasterbrick or vector of the minimum daily temperature}

\item{max_temp}{raster, rasterbrick or vector of the maximum daily temperature}

\item{avg_temp}{raster, rasterbrick or vector of the mean daily temperature}

\item{gdd_method}{string defining the method to use, "avg" for mean daily temperature or "be" for
the Baskerville-Emin method that fits a sine curve on the minimum and the maximum to account for daily
variation in temperature}

\item{top_temp}{value for the maximum temperature beyond growing degree day are not accumulating, default = NULL}
}
\description{
ggd_extract
Compute the growing degree day (GDD) using either the mean temperature or the Baskerville-Emin method,
with base temperature.
}
\details{
This function computes the Growing Degree Day on a specific base temperature and method (i.e., `avg` or `be`). The funciton
is build on two internal functions, gdd_avg_r() and gdd_be_r() to apply the specific gdd calculation method. The function
works on raster, rasterbrick (multiple layers), vector or matrix.
}
\author{
Reto Schmucki
}
