% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingDegreeDay.R
\name{get_date}
\alias{get_date}
\title{get_date}
\usage{
get_date(x, pattern, date_format)
}
\arguments{
\item{x}{rasterBrick or a vector of dates, from which to extract the dates of the time-series.}

\item{pattern}{string or characters to remove from the date name (remove the character "X" from "X2001.07.15")}

\item{date_format}{format of the date}
}
\description{
Extract a series of dates from the layers named in a rasterBrick or a vector of strings containing dates.
}
\details{
internal function to retrieve data from names.
}
\author{
Reto Schmucki
}
