% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingDegreeDay.R
\name{get_layer_indice}
\alias{get_layer_indice}
\title{get_layer_indice}
\usage{
get_layer_indice(
  x = NULL,
  pattern = NULL,
  date_format = "\%Y-\%m-\%d",
  indice_level = "year",
  week_start = getOption("lubridate.week.start", 1)
)
}
\arguments{
\item{x}{SpatRaster or a vector of dates, from which to extract the dates of the time-series.}

\item{pattern}{string or characters to remove from the date name (e.g. remove the character "X" from "X2001.07.15"), default is null.}

\item{date_format}{format of the date}

\item{indice_level}{string to define the time-period to use to define the indices, "year" or "month".}

\item{week_start}{option of lubridate starting day of the week, where 1 is Monday and 7 is Sunday.}
}
\description{
Compute a vector of indices of "year" or "month" based on the date extracted from the SpatRaster or a vector of
strings over the time period.
}
\details{
This function generate a vector of indices dividing the time-series in years or months. This indices are use
for the calculating the cummulative sum over years or month (see function cumsum_rb()).
}
\author{
Reto Schmucki
}
