% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{get_near_nona}
\alias{get_near_nona}
\title{get_near_nona}
\usage{
get_near_nona(x = x, y = y, x_cell = NULL)
}
\arguments{
\item{x}{raster object (raster, stack or brick)}

\item{y}{point as sf or spatial ojbect}

\item{x_cell}{cell id if already extracted, if not provide this will be computed}
}
\description{
Return cell number of the nearest data point available in the the raster layer
}
\details{
Linking a raster cell id to each points and searching for the nearest cell with data if points are within a NA cell
}
\author{
Reto Schmucki
}
