% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{get_thepoint}
\alias{get_thepoint}
\title{get_thepoint Function}
\usage{
get_thepoint(x, nc_index)
}
\arguments{
\item{x}{object obtained from the temporal_mean() function containing the array of averaged values}

\item{nc_index}{spatial indices corresponding to the coordinates of the geographic points}
}
\description{
Function to retrieve values corresponding to geographic points get_thepoint(data_nc$value_array,nc_index[p,])
}
\details{
This function is internal and used to extract value from the grid with the function point_grid_extract()
}
\author{
Reto Schmucki
}

