% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{point_grid_extract}
\alias{point_grid_extract}
\title{point_grid_extract}
\usage{
point_grid_extract(x, point_coord)
}
\arguments{
\item{x}{object obtained from the temporal_mean()  or temporal_sum()}

\item{point_coord}{sf object or a data.frame with coordinates in degree decimal
in two columns named "longitude" and "latitude" (epsg projection 4326 - wgs 84)}
}
\description{
Extract the mean climatic value (annual mean, monthly mean, sliding window 
mean) for each geographic points
}
\author{
Reto Schmucki
}
