% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{temporal_aggregate}
\alias{temporal_aggregate}
\title{temporal_aggregate}
\usage{
temporal_aggregate(
  x,
  y = NULL,
  agg_function = "mean",
  variable_name = "average temp",
  time_step = c("annual", "monthly")
)
}
\arguments{
\item{x}{object obtained from the extrat_nc_value function or a 
rasterBrick with dates as layers}

\item{y}{point to extract values, must be an sf or spatialPointDataFrame 
object (sp)}

\item{agg_function}{function used to aggregate data, "mean" or "sum"}

\item{variable_name}{character string to identify the resulting variable}

\item{time_step}{character string defining the level of averaging, "annual" 
or "monthly"}
}
\description{
Compute mean climatic values for specific periods from an object obtained 
from the extrat_nc_value function
}
\details{
The output is either a multilayer raster if no points are supplied 
in y or a data.table with aggregated statistic computed for each point 
provided in y. If points provided fall in areas with no data (e.g. in the sea
along the coast), the function will search for the nearest value available, up
to 3 cell in each direction. A distance (m) between the point and the raster 
cell used to retrieve the climate metric is calculated and documented in the 
data output. If the point is within a cell with value, the distance is set to
NA.
}
\author{
Reto Schmucki
}
