% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{temporal_mean}
\alias{temporal_mean}
\title{temporal_mean}
\usage{
temporal_mean(
  data_nc,
  time_avg = c("annual", "monthly", "window"),
  win_length = 30
)
}
\arguments{
\item{data_nc}{object obtained from the extrat_nc_value() function 
corresponding to the time-period of interest}

\item{time_avg}{character string defining the level of averaging, "annual", 
"monthly", "window"}

\item{win_length}{the number of days defining the span of the sliding window,
default set to 30 days}
}
\description{
Compute mean climatic value for specific periods "annual", "monthly", or 
using a sliding "window" of specific length, from an object obtained from 
the extrat_nc_value() function
}
\details{
Note that this function can be relatively slow if you compute 
sliding window average over a long period
}
\author{
Reto Schmucki
}
