% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{temporal_sum}
\alias{temporal_sum}
\title{temporal_sum}
\usage{
temporal_sum(data_nc, time_sum = c("annual", "monthly", "window"),
  win_length = 30)
}
\arguments{
\item{data_nc}{object obtained from the extrat_nc_value() function corresponding to the time-period of interest}

\item{time_sum}{character string defining the level of summation, "annual", "monthly", "window"}

\item{win_length}{the number of days defining the span of the sliding window, default set to 30 days}
}
\description{
This function compute sum of climatic value for specific periods "annual", "monthly", or using a sliding "window" of specific length, from and object obtained from the extrat_nc_value() function
}
\details{
This function can be relatively slow if you compute sliding window total over a long period
}
\author{
Reto Schmucki
}

