% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrsi_data_extract.R
\name{wrsi_data_extract}
\alias{wrsi_data_extract}
\title{Download WRSI data from https://edcftp.cr.usgs.gov/project/fews/dekadal/.}
\usage{
wrsi_data_extract(year_start, year_end, month_start, month_end, region, path)
}
\arguments{
\item{year_start}{start year.}

\item{year_end}{end year.}

\item{month_start}{start month.}

\item{month_end}{end month.}

\item{path}{path in which downloaded data will be stored.}
}
\value{
\code{wrsi_data_extract()} download all the GEOTIFF files for the period between year_start and year_end from https://edcftp.cr.usgs.gov/project/fews/dekadal.
}
\description{
Download WRSI data from https://edcftp.cr.usgs.gov/project/fews/dekadal/.
}
\examples{
\dontrun{
wrsi_data_extract (year_start = 20,year_end = 22,month_start=13,month_end=33, region="west_africa", path="./data-clean")
}
}
