% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps_data_extract.R
\name{chirps_data_extract}
\alias{chirps_data_extract}
\title{CHIRPS data download.}
\usage{
chirps_data_extract(year, month, path)
}
\arguments{
\item{year}{Year.}

\item{month}{Month.}

\item{path}{path in which downloaded data will be stored.}
}
\value{
\code{chirps_data_extract()} download all the GEOTIFF files for the period between year_start and year_end from https://data.chc.ucsb.edu/products/CHIRPS-2.0/global_monthly/tifs/.
}
\description{
CHIRPS data download.
}
\examples{
\dontrun{
path <- "C:/Users/WB495141/OneDrive - WBG/poverty/data/global/chirps"
y <- 1981
m <- "01"
chirps_data_extract(year_start = 20,year_end = 22,month_start=13,month_end=33, region="west_africa", path="./data-clean")
}
}
