% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{climateExtract}
\alias{climateExtract-package}
\alias{climateExtract}
\title{Extract Climate Data From a Local or online NETCDF}
\description{
A toolbox to extract and wrangle daily gridded climate data from NETCDF objects available from the European Climate Assessement & Datasets (ECAD) 
for specific time period, compute summarized values and link them to set of spatial points.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://retoschmucki.github.io/climateExtract}
  \item Report bugs at \url{https://github.com/RetoSchmucki/climateExtract/issues}
}

}
\author{
Reto Schmucki
}
\keyword{internal}
