% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_lc_data_extract.R
\name{esri_lc_data_extract}
\alias{esri_lc_data_extract}
\title{ESRI landcover data download.}
\usage{
esri_lc_data_extract(year, utm_id, path)
}
\arguments{
\item{year}{Year.}

\item{utm_id}{UTM ID.}

\item{path}{path in which downloaded data will be stored.}
}
\value{
\code{esri_lc_data_extract()} download GEOTIFF tile (utm_id) from https://livingatlas.arcgis.com/landcoverexplorer/#mapCenter=-121.64\%2C39.371\%2C11&mode=step&timeExtent=2017\%2C2022&year=2022.
}
\description{
ESRI landcover data download.
}
\examples{
\dontrun{
path <- "C:/Users/WB495141/OneDrive - WBG/poverty/data/global/esri_landcover"
year <- 2017
utm_id <- "36N"
esri_lc_data_extract(year,utm_id,path)
}
}
