% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaez_data_extract.R
\name{gaez_data_extract}
\alias{gaez_data_extract}
\title{GAEZ data download.}
\usage{
gaez_data_extract(path, res, model, rcp, time, indicator, crop, water_supply)
}
\arguments{
\item{res}{Resolution.}

\item{model}{Climate model.}

\item{rcp}{RCP.}

\item{time}{Time.}

\item{indicator}{Indicator.}

\item{crop}{Crop.}

\item{water_supply}{water supply.}
}
\value{
\code{download_wrsi()} download all the GEOTIFF files for the period between year_start and year_end from https://edcftp.cr.usgs.gov/project/fews/dekadal.
}
\description{
GAEZ data download.
}
\examples{
\dontrun{
gaez_data_extract(path, model, rcp, time, indicator, crop)
}
}
