% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaez_data_extract.R
\name{gaez_data_extract}
\alias{gaez_data_extract}
\title{GAEZ data download.}
\usage{
gaez_data_extract(path, model, rcp, time, indicator, crop)
}
\arguments{
\item{model}{Climate model.}

\item{rcp}{RCP.}

\item{time}{Time.}

\item{indicator}{Indicator.}

\item{crop}{Crop.}
}
\value{
\code{download_wrsi()} download all the GEOTIFF files for the period between year_start and year_end from https://edcftp.cr.usgs.gov/project/fews/dekadal.
}
\description{
GAEZ data download.
}
\examples{
\dontrun{
download_wrsi(year_start = 20,year_end = 22,month_start=13,month_end=33, region="west_africa", path="./data-clean")
}
}
