% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingDegreeDay.R
\name{get_layer_indice}
\alias{get_layer_indice}
\title{get_layer_indice}
\usage{
get_layer_indice(
  x = NULL,
  pattern = "X",
  date_format = "\%Y.\%m.\%d",
  indice_level = "year"
)
}
\arguments{
\item{x}{rasterBrick or a vector of dates, from which to extract the dates of the time-series.}

\item{pattern}{string or characters to remove from the date name (e.g. remove the character "X" from "X2001.07.15")}

\item{date_format}{format of the date}

\item{indice_level}{string to define the time-period to use to define the indices, "year" or "month".}
}
\description{
Compute a vector of indices of "year" or "month" based on the date extracted from the rasterBrick or a vector of
strings over the time period.
}
\details{
This function generate a vector of indices dividing the time-series in years or months. This indices are use
for the calculating the cummulative sum over years or month (see function cumsum_rb()).
}
\author{
Reto Schmucki
}
