% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{point_grid_extract}
\alias{point_grid_extract}
\title{point_grid_extract}
\usage{
point_grid_extract(data_nc, point_coord)
}
\arguments{
\item{data_nc}{object obtained from the temporal_mean() function corresponding 
to the period of interest and the level of averaging}

\item{point_coord}{a data.frame with three column named "site_id", "longitude",
 and "latitude" where coordinate of the points are in degree decimal 
(epsg projection 4326 - wgs 84)}
}
\description{
Extract the mean climatic value (annual mean, monthly mean, sliding window 
mean) for each geographic points
}
\author{
Reto Schmucki
}
