% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateExtract.R
\name{temporal_aggregate}
\alias{temporal_aggregate}
\title{temporal_aggregate}
\usage{
temporal_aggregate(
  x,
  y = NULL,
  agg_function = "mean",
  variable_name = "average temp",
  time_step = c("annual", "monthly")
)
}
\arguments{
\item{x}{object obtained from the extrat_nc_value() function or rasterBrick with date layers}

\item{y}{point to extract values, must be a sf or spatialPointDataFrame object (sp)}

\item{agg_function}{function used to aggregate data, "mean" or "sum"}

\item{variable_name}{character string to identify the resulting variable}

\item{time_step}{character string defining the level of averaging, "annual" or "monthly"}
}
\description{
This function compute mean climatic value for specific periods "annual", "monthly", or using a sliding "window" of specific length, from an object obtained from the extrat_nc_value() function
}
\details{
the output of this function is a multilayer raster if no point are supplied in y or a data.table when points are provided in y
}
\author{
Reto Schmucki
}
